Chnages

+++ src/components/Telegram/Telegram.service.ts
@@ -84,7 +84,7 @@ export class TelegramService implements OnModuleDestroy {
         }
         if (!this.hasClient(mobile)) {
             const telegramManager = new TelegramManager(user.session, user.mobile);
-            let client : TelegramClient
+            let client: TelegramClient
             try {
                 client = await telegramManager.createClient(handler);
                 await client.getMe();
@@ -115,7 +115,9 @@ export class TelegramService implements OnModuleDestroy {
                 }
             } catch (error) {
                 console.log("Parsing Error");
-                await client.destroy()
+                if (client) {
+                    await client.destroy()
+                }
                 const errorDetails = parseError(error);
                 if (contains(errorDetails.message.toLowerCase(), ['expired', 'unregistered', 'deactivated', "session_revoked", "user_deactivated_ban"])) {
                     console.log("Deleting User: ", user.mobile);
